% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_format.R
\name{case_format}
\alias{case_format}
\title{disaggregate case_age files}
\usage{
case_format(case_aggregated)
}
\value{
disaggregate data in "data/ecdc/age_cases.csv"
}
\description{
disaggregate case_age files
}
