% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_mutation.R
\name{create_table_mutation}
\alias{create_table_mutation}
\title{create_table_mutation}
\usage{
create_table_mutation(nextclade_data)
}
\arguments{
\item{nextclade_data}{}
}
\description{
create_table_mutation
}
