% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depletion_variant.R
\name{depletion_variant}
\alias{depletion_variant}
\title{The function increases the number of convictions of a variant over an age group. The increase is managed by the multiplicateur input (example 1.3 = 30\% less)}
\usage{
depletion_variant(data_aggregated, group, variants, multiplicateur)
}
\arguments{
\item{multiplicateur}{}
}
\value{

}
\description{
The function increases the number of convictions of a variant over an age group. The increase is managed by the multiplicateur input (example 1.3 = 30\% less)
}
