% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distortion_enrichment.R
\name{distortion_enrichment}
\alias{distortion_enrichment}
\title{Title}
\usage{
distortion_enrichment(data_aggregated, group, variants)
}
\arguments{
\item{variant}{}
}
\value{

}
\description{
Title
}
