% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_ECDC.R
\name{Download_ECDC}
\alias{Download_ECDC}
\title{Upload the files(variant, age_case, vaccination and hospitalisation) to the ecdc}
\usage{
Download_ECDC()
}
\value{
Upload the files in data/ecdc
}
\description{
Upload the files(variant, age_case, vaccination and hospitalisation) to the ecdc
}
