% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_genomic_data.R
\name{download_genomic_data}
\alias{download_genomic_data}
\title{Title}
\usage{
download_genomic_data(ncbitools.dir, nextclade.data.dir, nextclade.exfile)
}
\arguments{
\item{ncbitools.dir}{: : suppose que datasets et dataformat soit dans le meme dossier}

\item{nextclade.data.dir}{: dossier nextclade avec tous les fichiers de paramètres pour Sars-Cov-2}

\item{nextclade.exfile}{: fichier executable nextclade}
}
\description{
Title
}
