% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environnement.R
\name{environnement}
\alias{environnement}
\title{La fonction génére un environnement conda pour l'utilisation des outils nextclade et ncbi datasets}
\usage{
environnement()
}
\description{
La fonction génére un environnement conda pour l'utilisation des outils nextclade et ncbi datasets
}
