% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_booster_rates.R
\name{format_booster_rates}
\alias{format_booster_rates}
\title{Title}
\usage{
format_booster_rates(vaccination_data_formatted)
}
\arguments{
\item{vaccination_data_formatted}{}
}
\value{

}
\description{
Title
}
