% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_case.R
\name{format_case}
\alias{format_case}
\title{Formatting the files from ecdc :"https://opendata.ecdc.europa.eu/covid19/agecasesnational/csv/data.csv"- and only keep usefull columns}
\usage{
format_case(case_aggregated)
}
\value{

}
\description{
Formatting the files from ecdc :"https://opendata.ecdc.europa.eu/covid19/agecasesnational/csv/data.csv"- and only keep usefull columns
}
