% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_daily.R
\name{format_daily}
\alias{format_daily}
\title{Formatting the files daily from RIVM excel sheet - and only keep usefull columns}
\usage{
format_daily(daily, local_region, case_RIVM_formated)
}
\arguments{
\item{case_RIVM}{}
}
\value{

}
\description{
Formatting the files daily from RIVM excel sheet - and only keep usefull columns
}
