% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_hospitalisation.R
\name{format_hospitalisation}
\alias{format_hospitalisation}
\title{disaggregate hospitalisation files}
\usage{
format_hospitalisation(hospitalisation, vaccination, case_aggregated)
}
\arguments{
\item{hospitalisation}{}

\item{case_aggregated}{}
}
\value{

}
\description{
disaggregate hospitalisation files
}
