% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rivm.R
\name{format_rivm}
\alias{format_rivm}
\title{Formatting the files cases from RIVM excel sheet - and only keep usefull columns}
\usage{
format_rivm(case_RIVM, local_region)
}
\arguments{
\item{voc}{}
}
\value{

}
\description{
Formatting the files cases from RIVM excel sheet - and only keep usefull columns
}
