% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vaccination_data.R
\name{format_vaccination_data}
\alias{format_vaccination_data}
\title{Title}
\usage{
format_vaccination_data(ecdc_vaccination_data)
}
\arguments{
\item{ecdc_vaccination_data}{}
}
\value{

}
\description{
Title
}
