% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vaccines_rates.R
\name{format_vaccines_rates}
\alias{format_vaccines_rates}
\title{Title}
\usage{
format_vaccines_rates(vaccination_data_formatted)
}
\arguments{
\item{vaccination_data_formatted}{}
}
\value{

}
\description{
Title
}
