% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_variant.R
\name{format_variant}
\alias{format_variant}
\title{Formatting the files from ecdc :"https://opendata.ecdc.europa.eu/covid19/virusvariant/csv/data.csv"- and only keep usefull columns}
\usage{
format_variant(variants_aggregated, datedepart, datefin)
}
\value{

}
\description{
Formatting the files from ecdc :"https://opendata.ecdc.europa.eu/covid19/virusvariant/csv/data.csv"- and only keep usefull columns
}
