% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fxvariant}
\alias{fxvariant}
\title{Data set from the ECDC age group}
\format{
A data frame :
\describe{
\item{time}{The day of the year in  = "\%Y-\%m-\%d"}
\item{cases}{The number of cases}
\item{variant}{covid variant}
}
}
\usage{
data(fxvariant)
}
\description{
Data set from the ECDC. The dataset contains the data of Belgium between the 20th week and the 32nd week of 2021. The dataset has been formatted to be inserted in the simulator function.
}
\keyword{datasets}
