% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospitalisation_formatted_BE}
\alias{hospitalisation_formatted_BE}
\title{Data set from the ECDC hospitalisation}
\format{
A data frame :
\describe{
\item{country_code}{country_code of countries}
\item{year_week}{The year and the week}
\item{time}{The day of the year in  = "\%Y-\%m-\%d"}
\item{new_cases}{The number of cases}
}
}
\source{
\url{https://opendata.ecdc.europa.eu/covid19/hospitalicuadmissionrates/csv/data.csv}
}
\usage{
data(hospitalisation_formatted_BE)
}
\description{
Data set from the ECDC. The dataset contains the data of Belgium between the 20th week and the 32nd week of 2021. The dataset has been formatted to be inserted in the simulator function.
}
\keyword{datasets}
