% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_regression.R
\name{mutation_regression}
\alias{mutation_regression}
\title{Title}
\usage{
mutation_regression(data, numero_mutation, mutation, dateEnd, var_names_time)
}
\arguments{
\item{mutation}{}
}
\description{
Title
}
