% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{Prediction}
\alias{Prediction}
\title{Simulator: Predict the variant}
\usage{
Prediction(data, start = "2021-01", end = "2022-01")
}
\arguments{
\item{end}{}
}
\value{

}
\description{
Simulator: Predict the variant
}
\examples{


}
