% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_mutation.R
\name{select_mutation}
\alias{select_mutation}
\title{Title}
\usage{
select_mutation(
  data,
  dateStart,
  dateEnd,
  var_names_count,
  var_names_time,
  number
)
}
\arguments{
\item{number}{}
}
\description{
Title
}
