% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{Simulator: Predict the variant from a training set with a column time and a column variant to a testset with a column time}
\usage{
simulator(
  trainset,
  testset,
  start = "2021-01-01",
  end = "2022-01-01",
  country_code,
  outcome = "variant",
  desagregation = NULL
)
}
\arguments{
\item{trainset}{the data with a column year_week and variant}

\item{testset}{the data where the variant will be predicted. Need a year_week column}

\item{start}{the first week for the prediction}

\item{end}{the last week for the prediction}

\item{country_code}{a list with the countries you want to predict. In the form of country codes. (example: Belgium = BE)}

\item{desagregation}{}
}
\value{

}
\description{
Simulator: Predict the variant from a training set with a column time and a column variant to a testset with a column time
}
\examples{


}
