% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_fragmentation.R
\name{simulator_fragmentation}
\alias{simulator_fragmentation}
\title{Simulator: Predict the variant from a training set with a column time and a column variant to a testset with a column time}
\usage{
simulator_fragmentation(
  trainset,
  testset,
  time,
  geolocalisation,
  outcome,
  count = NULL,
  factor
)
}
\arguments{
\item{trainset}{the data with a column year_week and variant}

\item{testset}{the data where the variant will be predicted. Need a year_week column}

\item{factor}{}
}
\value{

}
\description{
Simulator: Predict the variant from a training set with a column time and a column variant to a testset with a column time
}
\examples{


}
