% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_genomic.R
\name{simulator_genomic}
\alias{simulator_genomic}
\title{Title}
\usage{
simulator_genomic(
  trainset,
  testset,
  time,
  geolocalisation,
  outcome,
  count,
  factor,
  bymonth = T,
  genomic_data,
  numberMutation,
  dateStart,
  dateEnd
)
}
\arguments{
\item{numberMutation}{}
}
\description{
Title
}
