% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_running.R
\name{simulator_running}
\alias{simulator_running}
\title{Title}
\usage{
simulator_running(
  trainset,
  testset,
  time,
  geolocalisation,
  outcome,
  count = NULL,
  factor
)
}
\arguments{
\item{factor}{}
}
\value{

}
\description{
Title
}
