% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_variant.R
\name{simulator_variant}
\alias{simulator_variant}
\title{Simulator: Predict the variant}
\usage{
simulator_variant(
  trainset,
  testset,
  start = "2021-01",
  end = "2022-01",
  country_code
)
}
\arguments{
\item{country_code}{}
}
\value{

}
\description{
Simulator: Predict the variant
}
\examples{


}
