% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_hospitalisation.R
\name{Table_hospitalisation}
\alias{Table_hospitalisation}
\title{disaggregate hospitalisation files}
\usage{
Table_hospitalisation()
}
\value{
disaggregate data in "data/ecdc/hospitalisation.csv"
}
\description{
disaggregate hospitalisation files
}
