% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_ecdc.R
\name{upload_ecdc}
\alias{upload_ecdc}
\title{Upload the files(variant, age_case) to the ecdc}
\usage{
upload_ecdc()
}
\value{
Upload the files in data/ecdc
}
\description{
Upload the files(variant, age_case) to the ecdc
}
