% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination_simulator.R
\name{vaccination_simulator}
\alias{vaccination_simulator}
\title{Title}
\usage{
vaccination_simulator(
  case_variants_aggregated,
  vaccination_data_formatted,
  vaccine_rates,
  booster_rates,
  start_date = NULL,
  end_date = NULL,
  country = NULL,
  vaccination_correction_factors = NULL
)
}
\arguments{
\item{vaccination_correction_factors}{}
}
\value{

}
\description{
Title
}
