% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_format.R
\name{variant_format}
\alias{variant_format}
\title{disaggregate variant files}
\usage{
variant_format(variants)
}
\value{
disaggregate data in "data/alldata.csv"
}
\description{
disaggregate variant files
}
