% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_genomic_data.R
\name{download_genomic_data}
\alias{download_genomic_data}
\title{Title}
\usage{
download_genomic_data(
  ecdcdata,
  var_names_variant,
  path_dataformat,
  path_dataset,
  path_nextclade
)
}
\arguments{
\item{path_nextclade}{}
}
\description{
Title
}
