% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_case.R
\name{format_case}
\alias{format_case}
\title{put the files case in the right format}
\usage{
format_case(case_aggregated)
}
\value{
disaggregate data in "data/ecdc/age_cases.csv"
}
\description{
put the files case in the right format
}
