% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_variant.R
\name{format_variant}
\alias{format_variant}
\title{put the files variant in the right format}
\usage{
format_variant(variants_aggregated)
}
\value{
disaggregate data in "data/alldata.csv"
}
\description{
put the files variant in the right format
}
